requirements.txt
requests
kafka-python
python-dotenv
pyspark
delta-spark